Copied props inside CodeWithCopyButton and replaced ! with ?
